fix: int to long for primary key, add ObjectWithNameDto for add director and fix DirectorRepository query string

+++ src/main/java/ru/yandex/practicum/filmorate/dal/FilmRepository.java
@@ -104,7 +104,7 @@ public class 	FilmRepository extends BaseRepository<Film> {
 		simpleInsert(INSERT_GENRE_QUERY, filmId, genreId);
 	}
 
-	public void addDirector(Long filmId, Integer directorId) {
+	public void addDirector(Long filmId, long directorId) {
 		simpleInsert(INSERT_DIRECTOR_QUERY, filmId, directorId);
 	}
 }
